Create a property wrapper to deal with json decoding.
Simple json decoding doesn't work well when we need to transform certain field from its original value.
For example, we receive a [SessionModel] array to represent an events sessions, but we want to store sessions in a more structured 2d array [[SessionModel]].

The standard way of doing this is customizing the init(decoder) initializer. However, this will bring a lot of boilerplate (multiple line of  try container.decode).
This property wrapper is created for this use case to avoid the above situation, and gives out a nicer structure (or class) declaration.